Implement createPriorityFeesInstruction function
