+++ src/transactions/constants.ts
@@ -23,3 +23,5 @@ export enum BorrowType {
   CNft = 'CNft',
   Default = 'Default',
 }
+
+export const DEFAULT_PRIORITY_FEE = 300_000 //? Micro lamports
