+++ src/transactions/helpers/index.ts
@@ -1,4 +1,4 @@
-export * from './common'
+export * from './createPriorityFeesInstruction'
 export * from './enqueueErrorSnackbar'
 export * from './getTxnErrorDefinition'
 export * from './signAndConfirmTransaction'
