+++ src/transactions/loans/makeClaimLoanTransaction.ts
@@ -13,7 +13,7 @@ import { BONDS } from '@banx/constants'
 import { sendTxnPlaceHolder } from '@banx/utils'
 
 import { fetchRuleset } from '../functions'
-import { createInstructionsWithPriorityFees } from '../helpers'
+import { createPriorityFeesInstruction } from '../helpers'
 
 export type MakeClaimActionParams = {
   loan: Loan
@@ -54,13 +54,10 @@ export const makeClaimAction: MakeClaimAction = async (ixnParams, { connection, 
       bondTradeTransaction: optimisticResult.bondTradeTransaction,
     }
 
-    const instructionsWithPriorityFees = await createInstructionsWithPriorityFees(
-      instructions,
-      connection,
-    )
+    const priorityFeeInstruction = await createPriorityFeesInstruction(instructions, connection)
 
     return {
-      instructions: instructionsWithPriorityFees,
+      instructions: [...instructions, priorityFeeInstruction],
       signers,
       result: optimisticLoan,
       lookupTables: [new web3.PublicKey(LOOKUP_TABLE)],
@@ -101,14 +98,10 @@ export const makeClaimAction: MakeClaimAction = async (ixnParams, { connection, 
       fraktBond: optimisticResult.fraktBond,
       bondTradeTransaction: optimisticResult.bondTradeTransaction,
     }
-
-    const instructionsWithPriorityFees = await createInstructionsWithPriorityFees(
-      instructions,
-      connection,
-    )
+    const priorityFeeInstruction = await createPriorityFeesInstruction(instructions, connection)
 
     return {
-      instructions: instructionsWithPriorityFees,
+      instructions: [...instructions, priorityFeeInstruction],
       signers,
       result: optimisticLoan,
       lookupTables: [new web3.PublicKey(LOOKUP_TABLE)],
