+++ src/transactions/staking/stakeBanxToken.ts
@@ -7,7 +7,7 @@ import { CreateTransactionDataFn } from 'solana-transactions-executor'
 import { BONDS } from '@banx/constants'
 import { sendTxnPlaceHolder } from '@banx/utils'
 
-import { createInstructionsWithPriorityFees } from '../helpers'
+import { createPriorityFeesInstruction } from '../helpers'
 
 export type StakeBanxTokenActionParams = {
   tokensToStake: BN
@@ -35,13 +35,10 @@ export const stakeBanxTokenAction: StakeBanxTokenAction = async (
 
   const { instructions, signers } = await stakeBanxToken(params)
 
-  const instructionsWithPriorityFees = await createInstructionsWithPriorityFees(
-    instructions,
-    connection,
-  )
+  const priorityFeeInstruction = await createPriorityFeesInstruction(instructions, connection)
 
   return {
-    instructions: instructionsWithPriorityFees,
+    instructions: [...instructions, priorityFeeInstruction],
     signers,
     lookupTables: [],
   }
