+++ src/transactions/staking/subscribeBanxAdventureAction.ts
@@ -5,7 +5,7 @@ import { CreateTransactionDataFn } from 'solana-transactions-executor'
 import { BONDS } from '@banx/constants'
 import { sendTxnPlaceHolder } from '@banx/utils'
 
-import { createInstructionsWithPriorityFees } from '../helpers'
+import { createPriorityFeesInstruction } from '../helpers'
 
 export type SubscribeBanxAdventureParams = {
   weeks: number[]
@@ -33,13 +33,10 @@ export const subscribeBanxAdventureAction: SubscribeBanxAdventureAction = async 
     sendTxn: sendTxnPlaceHolder,
   })
 
-  const instructionsWithPriorityFees = await createInstructionsWithPriorityFees(
-    instructions,
-    connection,
-  )
+  const priorityFeeInstruction = await createPriorityFeesInstruction(instructions, connection)
 
   return {
-    instructions: instructionsWithPriorityFees,
+    instructions: [...instructions, priorityFeeInstruction],
     signers: signers,
     lookupTables: [],
   }
