+++ src/utils/accounts/index.ts
@@ -1,6 +1,5 @@
 import { useConnection, useWallet } from '@solana/wallet-adapter-react'
 import { useQuery } from '@tanstack/react-query'
-import axios from 'axios'
 import { web3 } from 'fbonds-core'
 
 type UseNativeAccount = ({ isLive }: { isLive?: boolean }) => web3.AccountInfo<Buffer> | null
@@ -34,32 +33,4 @@ export const useSolanaBalance: UseSolanaBalance = ({ isLive = true } = {}) => {
   const balance = (account?.lamports || 0) / web3.LAMPORTS_PER_SOL
 
   return balance
-}
-
-export const calculatePriorityFees = async (
-  connection: web3.Connection,
-  accountKeys: string[],
-): Promise<number> => {
-  try {
-    const response = await axios.post(connection.rpcEndpoint, {
-      jsonrpc: '2.0',
-      id: 'my-id',
-      method: 'getPriorityFeeEstimate',
-      params: [
-        {
-          accountKeys,
-          options: {
-            includeAllPriorityFeeLevels: true,
-          },
-        },
-      ],
-    })
-
-    const { result } = response.data
-
-    return Math.trunc(result.priorityFeeLevels.veryHigh)
-  } catch (error) {
-    console.error('Error calculating priority fees:', error)
-    throw error
-  }
 }
